Configure: restored "no-threads" in OpenSSL builds.
This was previously used, but was incorrectly removed in b45ca139218f
while removing old threads remnants.  Instead of using it conditionally
when threads are not used, we now set in unconditionally, as even with
thread pools enabled we never call OpenSSL functions in threads.

This fixes resulting binary when using --with-openssl with OpenSSL 1.1.0+
and without -lpthread linked (notably on FreeBSD without PCRE).

--HG--
branch : nginx